server/requirements.txt
astral==1.10.1
cachetools==3.1.0
croniter==0.3.30
firebase-admin==2.16.0
Flask==1.0.2
google-api-python-client==1.7.8
numpy==1.16.2
oauth2client==4.1.3
Pillow==6.0.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.21.0
scipy==1.2.1
